Set the port to null if unset
If there are hosts (including a port) defined in ~.ssh/config it is common
to only us ethat host's name without a port with SSH. So allow to not
specify a port in Gerrie's config. Instead of not setting it, set it
explicitly to null so that bootstrapSSHDataService() do esnot stumple over
it complaining about an "Undefined index: port".